Add buf size assert to samples
The constructor is not hot so its assert never fails
the assert is a bit strict because if you know what you are doing you
could pass in a bigger buffer than needed by twbl and it could still
work.

But powerusers can just edit that line since its directly in the users
code anyways.